- Create separate placeholder components for MyInfo, Job Openings, Reports and Messages - Ensure React Router routes work..
