+++ client/src/components/JobOpenings.jsx
@@ -1,0 +1,13 @@
+import React, { Component } from 'react'
+
+ class JobOpenings extends Component {
+  render() {
+    return (
+      <div>
+        This is a placeholder.
+      </div>
+    )
+  }
+}
+
+export default JobOpenings;
