+++ client/src/components/Messages.jsx
@@ -1,0 +1,13 @@
+import React, { Component } from 'react'
+
+ class Messages extends Component {
+  render() {
+    return (
+      <div>
+        This is a placeholder.
+      </div>
+    )
+  }
+}
+
+export default Messages;
