+++ client/src/components/MyInfo.jsx
@@ -1,0 +1,13 @@
+import React, { Component } from 'react'
+
+ class MyInfo extends Component {
+  render() {
+    return (
+      <div>
+        This is a placeholder.
+      </div>
+    )
+  }
+}
+
+export default MyInfo;
