+++ client/src/components/Reports.jsx
@@ -1,0 +1,13 @@
+import React, { Component } from 'react'
+
+ class Reports extends Component {
+  render() {
+    return (
+      <div>
+        This is a placeholder.
+      </div>
+    )
+  }
+}
+
+export default Reports;
